.history/properties/models_20241016141824.py
from cloudinary.models import CloudinaryField
from cloudinary_storage.storage import VideoMediaCloudinaryStorage
from cloudinary_storage.validators import validate_video
from django.conf import settings
from django.contrib.auth.models import User
from django.db import models

