requirements.txt
asgiref==3.8.1
babel==2.16.0
black==24.8.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
cryptography==43.0.0
dj-database-url==2.2.0
dj3-cloudinary-storage==0.0.6
Django==5.1
django-allauth==64.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.9.2
django-extensions==3.2.3
django-phonenumber-field==8.0.0
flake8==7.1.1
gunicorn==23.0.0
idna==3.8
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.47
pillow==10.4.0
platformdirs==4.3.2
psycopg2==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
python-http-client==3.3.7
python-magic==0.4.27
requests==2.32.3
sendgrid==6.9.1
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
